style: refactor and move nextstep and close buttons
