+++ plugin-hrm-form/src/components/CSAMReport/CSAMReportStatusScreen.tsx
@@ -19,7 +19,8 @@ import React from 'react';
 import { Template } from '@twilio/flex-ui';
 import Close from '@material-ui/icons/Close';
 
-import { BottomButtonBar, Box, HeaderCloseButton, HiddenText, Row, StyledNextStepButton } from '../../styles/HrmStyles';
+import { BottomButtonBar, Box, HiddenText, Row } from '../../styles/HrmStyles';
+import { StyledNextStepButton, HeaderCloseButton } from '../../styles/buttons';
 import {
   BoldDescriptionText,
   ButtonText,
