+++ plugin-hrm-form/src/components/case/EditCaseSummary.tsx
@@ -30,9 +30,9 @@ import {
   Box,
   ColumnarBlock,
   Container,
-  StyledNextStepButton,
   TwoColumnLayout,
 } from '../../styles/HrmStyles';
+import { StyledNextStepButton } from '../../styles/buttons';
 import ActionHeader from './ActionHeader';
 import { RootState } from '../../states';
 import * as RoutingActions from '../../states/routing/actions';
