+++ plugin-hrm-form/src/components/caseMergingBanners/ContactRemovedFromCaseBanner.tsx
@@ -19,7 +19,8 @@ import { connect } from 'react-redux';
 import { Template } from '@twilio/flex-ui';
 import { Close } from '@material-ui/icons';
 
-import { HeaderCloseButton, HiddenText } from '../../styles/HrmStyles';
+import { HiddenText } from '../../styles/HrmStyles';
+import { HeaderCloseButton } from '../../styles/buttons';
 import selectContactByTaskSid from '../../states/contacts/selectContactByTaskSid';
 import WarningIcon from './WarningIcon';
 import { BannerContainer, Text } from './styles';
