+++ plugin-hrm-form/src/components/common/forms/DownloadFile.tsx
@@ -20,7 +20,8 @@ import DownloadIcon from '@material-ui/icons/GetApp';
 import { Template } from '@twilio/flex-ui';
 
 import { getHrmConfig } from '../../../hrmConfig';
-import { Flex, Box, StyledNextStepButton, HiddenText } from '../../../styles/HrmStyles';
+import { Flex, Box, HiddenText } from '../../../styles/HrmStyles';
+import { StyledNextStepButton } from '../../../styles/buttons';
 import { formatFileNameAtAws } from '../../../utils';
 import { fetchHrmApi, generateSignedURLPath } from '../../../services/fetchHrmApi';
 import { FormTargetObject } from './types';
