+++ plugin-hrm-form/src/components/common/forms/UploadFileInput.tsx
@@ -21,15 +21,8 @@ import { Template } from '@twilio/flex-ui';
 import AttachFileIcon from '@material-ui/icons/AttachFile';
 import { CircularProgress } from '@material-ui/core';
 
-import {
-  Box,
-  FormError,
-  Row,
-  FormInput,
-  UploadFileLabel,
-  UploadFileFileName,
-  StyledNextStepButton,
-} from '../../../styles/HrmStyles';
+import { Box, FormError, Row, FormInput, UploadFileLabel, UploadFileFileName } from '../../../styles/HrmStyles';
+import { StyledNextStepButton } from '../../../styles/buttons';
 import { StyledLink } from '../../../styles/search';
 import UploadIcon from '../icons/UploadIcon';
 import { formatFileNameAtAws } from '../../../utils';
