+++ plugin-hrm-form/src/components/profile/ProfileIdentifierBanner/ProfileIdentifierBanner.tsx
@@ -20,7 +20,7 @@ import { connect, ConnectedProps } from 'react-redux';
 import { Template } from '@twilio/flex-ui';
 
 import { useIdentifierByIdentifier, useProfileProperty } from '../../../states/profile/hooks';
-import { YellowBanner } from '../../../styles/previousContactsBanner';
+import { YellowBanner } from './styles';
 import { Bold } from '../../../styles/HrmStyles';
 import { StyledLink } from '../../../styles/search';
 import { ChannelTypes, channelTypes } from '../../../states/DomainConstants';
