+++ plugin-hrm-form/src/components/resources/resourceView/ExpandableSection.tsx
@@ -18,7 +18,7 @@ import ArrowRightTwoTone from '@material-ui/icons/ArrowRightTwoTone';
 import ArrowDropDownTwoTone from '@material-ui/icons/ArrowDropDownTwoTone';
 
 import { SectionCollapse } from '../../../styles/search';
-import { SectionTitleContainer, SectionTitleButton, SectionTitleText } from '../../../styles/ReferrableResources';
+import { SectionTitleContainer, SectionTitleButton, SectionTitleText } from '../styles';
 
 type ExpandableSectionProps = {
   title: string | JSX.Element;
