+++ plugin-hrm-form/src/components/resources/resourceView/ResourceAttribute.tsx
@@ -18,7 +18,7 @@ import React from 'react';
 import { Template } from '@twilio/flex-ui';
 
 import { Box, Column } from '../../../styles/HrmStyles';
-import { ResourceAttributeContent, ResourceAttributeDescription } from '../../../styles/ReferrableResources';
+import { ResourceAttributeContent, ResourceAttributeDescription } from '../styles';
 import ExpandableAttributeContent from './ExpandableAttributeContent';
 
 type Props = {
