+++ plugin-hrm-form/src/components/tabbedForms/AddCaseButton.tsx
@@ -20,7 +20,8 @@ import FolderIcon from '@material-ui/icons/CreateNewFolderOutlined';
 import KeyboardArrowUpIcon from '@material-ui/icons/KeyboardArrowUp';
 import KeyboardArrowDownIcon from '@material-ui/icons/KeyboardArrowDown';
 
-import { StyledAddNewCaseDropdown, StyledAddNewCaseDropdownList, StyledNextStepButton } from '../../styles/HrmStyles';
+import { StyledAddNewCaseDropdown, StyledAddNewCaseDropdownList } from './styles';
+import { StyledNextStepButton } from '../../styles/buttons';
 
 type OwnProps = {
   handleNewCaseType: () => void;
