+++ plugin-hrm-form/src/components/tabbedForms/BottomBar.tsx
@@ -21,13 +21,9 @@ import { CircularProgress } from '@material-ui/core';
 import FolderIcon from '@material-ui/icons/CreateNewFolderOutlined';
 import { DefinitionVersionId } from 'hrm-form-definitions';
 
-import {
-  AddedToCaseButton,
-  BottomButtonBar,
-  Box,
-  SaveAndEndButton,
-  StyledNextStepButton,
-} from '../../styles/HrmStyles';
+import { Box, BottomButtonBar } from '../../styles/HrmStyles';
+import { AddedToCaseButton } from './styles';
+import { StyledNextStepButton, SaveAndEndButton } from '../../styles/buttons';
 import * as RoutingActions from '../../states/routing/actions';
 import { completeTask } from '../../services/formSubmissionHelpers';
 import { hasTaskControl } from '../../utils/transfer';
