+++ plugin-hrm-form/src/styles/HrmStyles.tsx
@@ -26,8 +26,6 @@ import Tooltip from '@material-ui/core/Tooltip';
 
 import HrmTheme from './HrmTheme';
 
-export const BottomButtonBarHeight = 55;
-
 type BoxProps = {
   width?: string;
   height?: string;
@@ -94,27 +92,6 @@ export const Absolute = styled('div')<AbsoluteProps>`
   right: ${({ right }) => right || 'auto'};
 `;
 Absolute.displayName = 'Absolute';
-
-export const TabbedFormsContainer = styled('div')`
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-`;
-TabbedFormsContainer.displayName = 'TabbedFormsContainer';
-
-type TabbedFormTabContainerProps = {
-  display: boolean;
-};
-
-export const TabbedFormTabContainer = styled(({ display, ...rest }: TabbedFormTabContainerProps) => <Box {...rest} />)<
-  TabbedFormTabContainerProps
->`
-  padding: 32px 20px 12px 20px;
-  background-color: white;
-  display: ${({ display }) => (display ? 'block' : 'none')};
-  height: ${({ display }) => (display ? '100%' : '0px')};
-`;
-TabbedFormTabContainer.displayName = 'TabbedFormTabContainer';
 
 const containerLeftRightMargin = '5px';
 
@@ -274,92 +251,7 @@ export const StyledMenuItem = styled(MenuItem)`
 `;
 StyledMenuItem.displayName = 'StyledMenuItem';
 
-type StyledNextStepButtonProps = {
-  secondary?: string; // string to prevent console errors
-  disabled?: boolean;
-  margin?: string;
-};
-
-export const AddedToCaseButton = styled('p')`
-  display: flex;
-  align-items: center;
-  font-size: 13px;
-  font-weight: 800;
-  letter-spacing: normal;
-  color: ${HrmTheme.colors.secondaryButtonTextColor};
-  border: none;
-  padding: 4px 10px;
-  min-width: auto;
-`;
-AddedToCaseButton.displayName = 'AddedToCaseButton';
-
-export const SaveAndEndButton = styled(Button)<StyledNextStepButtonProps>`
-  display: flex;
-  align-items: center;
-  font-size: 13px;
-  font-weight: 800;
-  letter-spacing: normal;
-  color: ${HrmTheme.colors.buttonTextColor};
-  background: linear-gradient(to top, ${HrmTheme.colors.declineColor}, ${HrmTheme.colors.declineColor});
-  border: none;
-  padding: 4px 10px;
-  min-width: auto;
-  cursor: ${props => (props.disabled ? 'not-allowed' : 'pointer')};
-
-  &&:focus {
-    outline-style: auto;
-    outline-width: initial;
-  }
-
-  &&:active {
-    background-color: rgba(255, 255, 255, 0.3);
-  }
-`;
-SaveAndEndButton.displayName = 'SaveAndEndButton';
-
-export const StyledNextStepButton = styled(Button)<StyledNextStepButtonProps>`
-  display: flex;
-  align-items: center;
-  font-size: 13px;
-  letter-spacing: normal;
-  color: ${props =>
-    props.secondary?.toLowerCase() === 'true'
-      ? HrmTheme.colors.secondaryButtonTextColor
-      : HrmTheme.colors.buttonTextColor};
-  border: none;
-  border-radius: 4px;
-  margin: ${props => (props.margin ? props.margin : '0')};
-  padding: 4px 10px;
-  min-width: auto;
-  background-color: ${props =>
-    props.disabled
-      ? HrmTheme.colors.disabledColor
-      : props.secondary?.toLowerCase() === 'true'
-      ? HrmTheme.colors.secondaryButtonColor
-      : HrmTheme.colors.defaultButtonColor};
-  cursor: ${props => (props.disabled ? 'not-allowed' : 'pointer')};
-  ${p =>
-    getBackgroundWithHoverCSS(
-      p.disabled
-        ? HrmTheme.colors.base5
-        : p.secondary?.toLowerCase() === 'true'
-        ? HrmTheme.colors.secondaryButtonColor
-        : HrmTheme.colors.defaultButtonColor,
-      true,
-      false,
-      p.disabled,
-    )};
-
-  &&:focus {
-    outline-style: auto;
-    outline-width: initial;
-  }
-
-  &&:active {
-    background-color: rgba(255, 255, 255, 0.3);
-  }
-`;
-StyledNextStepButton.displayName = 'StyledNextStepButton';
+export const BottomButtonBarHeight = 55;
 
 export const BottomButtonBar = styled('div')`
   display: flex;
@@ -1181,127 +1073,6 @@ export const Bold = styled('span')`
 `;
 
 Bold.displayName = 'Bold';
-
-export const CSAMReportButtonText = styled(FontOpenSans)`
-  font-size: 12px;
-  color: ${HrmTheme.colors.hyperlinkColor};
-  font-weight: 600;
-`;
-CSAMReportButtonText.displayName = 'CSAMReportButtonText';
-
-const TabbedFormsHeaderButton = styled(ButtonBase)`
-  &:focus {
-    outline-color: ${HrmTheme.colors.focusColor};
-    outline-style: solid;
-    outline-width: medium;
-  }
-
-  &:hover {
-    background-color: #f2f2f2;
-    cursor: pointer;
-  }
-`;
-TabbedFormsHeaderButton.displayName = 'TabbedFormsHeaderButton';
-
-export const StyledBackButton = TabbedFormsHeaderButton;
-StyledBackButton.displayName = 'StyledBackButton';
-
-export const StyledCSAMReportButton = TabbedFormsHeaderButton;
-StyledCSAMReportButton.displayName = 'StyledCSAMReportButton';
-
-export const HeaderCloseButton = styled(ButtonBase)`
-  && {
-    margin-left: auto;
-  }
-
-  &:focus {
-    outline-color: ${HrmTheme.colors.focusColor};
-    outline-style: solid;
-    outline-width: medium;
-  }
-
-  &:hover {
-    background-color: #f2f2f2;
-    cursor: pointer;
-  }
-`;
-HeaderCloseButton.displayName = 'HeaderCloseButton';
-
-export const StyledCSAMReportDropdown = styled('ul')`
-  position: absolute;
-  right: 0;
-  left: auto;
-  box-shadow: 0 10px 15px -3px rgba(46, 41, 51, 0.08), 0 4px 6px -2px rgba(71, 63, 79, 0.16);
-  font-size: 0.875rem;
-  z-index: 9999;
-  min-width: 10rem;
-  padding: 10px 100px 10px 24px;
-  list-style: none;
-  background-color: #fff;
-  border-radius: 0 0 5px 5px;
-  margin-right: 20px;
-`;
-
-StyledCSAMReportDropdown.displayName = 'StyledCSAMReportDropdown';
-
-export const StyledCSAMReportDropdownList = styled('button')`
-  position: relative;
-  font-size: 14px;
-  display: block;
-  color: inherit;
-  min-width: 10rem;
-  width: 215px;
-  padding: 7px 0 7px 0;
-  margin: 0 -100px 0 -25px;
-  text-decoration: none;
-  &:hover {
-    background-color: #f2f2f2;
-    cursor: pointer;
-  }
-  background: none;
-  border: none;
-`;
-StyledCSAMReportDropdownList.displayName = 'StyledCSAMReportDropdownList';
-
-export const StyledAddNewCaseDropdown = styled('ul')`
-  position: absolute;
-  right: -12%;
-  bottom: 110%;
-  display: flex;
-  box-shadow: 0px 4px 16px 0px rgba(18, 28, 45, 0.2);
-  -webkit-box-shadow: 0px 4px 16px 0px rgba(18, 28, 45, 0.2);
-  -moz-box-shadow: 0px 4px 16px 0px rgba(18, 28, 45, 0.2);
-  font-size: 0.875rem;
-  z-index: 9999;
-  width: 164px;
-  padding: 10px 0 10px 0;
-  flex-direction: column;
-  align-items: flex-start;
-  background: var(--background-color-background-body, #fff);
-  border: 1px solid var(--border-color-border-weaker, #e1e3ea);
-  border-radius: 8px;
-  margin-right: 20px;
-`;
-StyledAddNewCaseDropdown.displayName = 'StyledAddNewCaseDropdown';
-
-export const StyledAddNewCaseDropdownList = styled('button')`
-  position: relative;
-  font-size: 14px;
-  display: flex;
-  color: inherit;
-  min-width: 10.1rem;
-  align-items: flex-start;
-  align-self: stretch;
-  padding: 7px 0 7px 18px;
-  text-decoration: none;
-  &:hover {
-    background-color: #f2f2f2;
-    cursor: pointer;
-  }
-  background: none;
-  border: none;
-`;
-StyledAddNewCaseDropdownList.displayName = 'StyledAddNewCaseDropdownList';
 
 export const TypingIndicatorText = styled(FontOpenSans)`
   font-size: 10px;
